requirements.txt
# ZmartBot Platform Requirements
# Python 3.11+ required

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
psycopg2-binary==2.9.9

# Security
slowapi==0.1.9
python-jose[cryptography]==3.3.0
hvac==2.3.0
bcrypt==4.1.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# API Clients
httpx==0.25.1
aiohttp==3.9.0
requests==2.31.0
websocket-client==1.6.4

# Trading APIs
ccxt==4.1.56
kucoin-python==1.0.11

# AI/ML
openai==1.3.5
langchain==0.0.340
tiktoken==0.5.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Monitoring
prometheus-client==0.19.0
influxdb-client==1.38.0

# Message Queue
pika==1.3.2
celery==5.3.4
kombu==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
croniter==2.0.1
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
isort==5.12.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Production
gunicorn==21.2.0
supervisor==4.2.5